The Indicator Update
This update includes indicators for if you're holding objects in either hands that work - however we've encountered a glitch. If you're holding an object in one hand and try to pick up an object in another, the object either is not picked up or is destroyed, and your empty arm becomes full (indicated by check mark).