pot drop sound
adds three sound effects for dropping an item in a pot. also the timeline cooking system works. there are 4 levels of steam that rise above the pot when cooking. the player can pull out a recipe at any of these times, and it will adjust the quality of the food based on the time it is pulled out.