Finalized Runes
I'm not completely sold on the size and designs of the runes I've made but the system should work without a hitch now!

Keep in mind every specific instance of oCookingVessel (such as oPot) is going to need its own draw event to draw the runes with the way I have it programmed right now.