Object Hierarchy Change and Sound Edit
Changed each instance of oPreparedIngredient to be a child of its respective oIngredient, and I'll explain why.

I did so because I added a new variable for each oHoldableObject called "pickUpSound," and changed the step event for every holdable to play that sound (right now there's a cheesy default sound for any object that doesn't get a pickUpSound assigned). This way the sounds could be uniform for every item, and it could be inherited by their children. This also cleaned up the oPreparedIngredient code a good bit.