New Target Type, Item, and Input
Added oPlayerTarget, which is counted as "hit" when it collides with the player.

Added oPebble. Technically it's an ingredient, but it won't break the game if you cook it. It's going to be used to hit the targets in the tutorial, but it'd be best to destroy all instances of it after that.

Made it so you can progress dialog with the jump button as well (this will be better for controller play). Currently the chef can still move during cutscenes but if that's removed this'll work like a charm.