Skipping Cutscenes WIP
Added in a few new sound tweaks and dialog tweaks (some of these are present in the current demo build on itch but I forgot to push them :-/)

Added a WIP system for skipping cutscenes. It currently doesn't really work and it's only in place for the first demo cutscene (not the tutorial). Stil not sure why it's not working yet. Probably something really simple.