More New Tutorial (WIP)
Added more to the tutorial until I hit a weird inventory bug. Also added in lines in oOldMan that allow him to repeat his words to you if you press E on him (this was a feature in the old cutscene, I just updated it). oNewTutorial already has a slew of bugs - details are in its create event. Check them out and if you can solve them pwease do :3