resolution fixes
i think i fixed the resolution scaling. press Z to zoom. go to oDisplayManager and change "ideal_height" to whatever value you want. this will be the resolution of the height of the game.

when you can no longer zoom without pixel distortion, the camera of the game will scale up to accommodate the empty space. the only problem with this is that youll be able to see more or less of the screen depending on the size of your monitor. but i dont think this is that much of an issue.