Movement Sounds WIP
Added two footstep placeholder sounds (free mp3 downloads and a script that plays the sounds as the player walks. Not sure if this is perfect yet - there seem to be some discrepancies I want to test out. Also had to add some booleans to the oPlayer create event because I couldn't figure out what else to do I'm sorry.