New Holding Sprites
Added in holding sprites for idle, moving, jumping, falling, and taking damage. This might not be the most efficient way to do it, but I've included a search through of the held_items list everywhere a sprite is set, and then there it is set accordingly if something is being held.

scr_movement_sounds was updated to still play sounds with these sprites as well (sorry to whoever commented that out - but the sounds are back in this push). One major issue is that while things are being held, the 1-way platforms don't work. I'm not sure why, like at all. Couldn't find anything in their code.