New Cooking System
Fixed problems with grid implementation of cooking! Now cooking works with a new grid system, meaning that we no longer need oIngredientHolder (hooray)! The benefit here is that the items are destroyed when they interact with the pot, and their values are stored in item_grid, which is referenced during cooking. This can also be referenced to take individual items out of the pot later.