Pot Holding Sprites Implemented
Player can now hold the pot and it looks right due to layering. The player's arms are drawn in oCookingVessel's draw event, which now also includes a script that tracks the offset of the big holding animations.

Is it just me, or is it a little high up? Maybe we should lower it some.