Recipe Sprite Layering and oRecipe Information Handling
Added in code that allows for Phillip's WIP sprites to be rendered when thrown (the chef still caries a large chunk of cheddar cheese - but we'll get there). 

To do this, I had to re-add in the code that stores an oRecipe's data when taken out of a pot and recreates it when thrown. Took me a little longer than it should have - but it's here!