Soup Hand Tracking and SQUIPPLE
Added in a little bit of code so that the soups track in the players hands and the arms are rendered appropriately (all right)!

GameMaker decided to get angry at me and not allow me to rename scr_offset_pot to scr_offset_hands, so for a while it's going to be that. Hopefully I remember to change it in the future so it's a little less confusing. 

Also, Spider Squirrel Legs are now known in-game as Squipple Legs. Tragically, their sprite names and object names haven't changed, because this ended up being a massive pain due to how many things have the words "spider" and "squirrel" in them. Still, it works! Fried squipple for everyone!