Hold E To Pick up Potty, Ingredients in Palette, Etc.
Added in oPlayerInput.key_interact_held as a new variable to see if players are holding the pick up button. Added this to potty. If players hold down E, they can pick up potty. Right now, a crudely animated bar shows the progress. Later, we should add in an E that shows it off somehow (be it filling up red, or maybe even a ring that fills because those always look good). 

Also messed around with Sub's ingredient sprites and made them fit our palette. I think overall they look a lot more natural to the game (maybe with the exception of the Fugi Nut).

Also added in preliminary fixes for how multiple arms are drawn when the chef holds a big item. Currently the fixes only work when he's facing right. Not sure why.