New Campfire Sprite
I added a really wild fire on top of Phillip's campfire (which is now in-palette). 

Also, the sprite for holdMePotty is in. It is not implemented in the game yet, but it WILL BE!!! THE DAY SHALL COME WHEN ALL WILL KNOW OF POTTY'S DESIRE TO BE HELD. THE MEN WILL LAUGH AND THE WOMEN WILL DANCE, DRINK AND MERRIMENT WILL OERTAKE THE LAND AND THE SCION OF THE FINAL DAYS SHALL BRING PEACE AND